refactored index.vue
